requirements.txt
Django==4.2.10
djangorestframework==3.14.0
psycopg2-binary==2.9.3
drf-yasg==1.21.7